fix warn: - Reports leading whitespaces before Go directives in comments. - Incorrect usages of fmt.Printf, fmt.Println, and similar formatting and printing functions. - Redundant parentheses in expressions and types. - Types in variable and constant declarations that can be omitted since they can be inferred by the compiler. Such types are redundant, omitting them may improve readability of the code. - Type conversions that may be omitted. - Aliases of imported packages that may be omitted.

+++ common/ledger/blkstorage/blockindex.go
@@ -390,7 +390,7 @@ func importTxIDsFromSnapshot(
 			return err
 		}
 		batch.Put(
-			constructTxIDKey(txID, lastBlockNumInSnapshot, uint64(i)),
+			constructTxIDKey(txID, lastBlockNumInSnapshot, i),
 			[]byte{},
 		)
 		if (i+1)%importTxIDsBatchSize == 0 {
